Add sanitizer configurations
Just double checking we're not failing any of the basic static analysis
provided by the sanitizers. Note that a few of these are only available
on linux.

* `--config asan` - Address sanitizer

* `--config tsan` - Thread sanitizer

* `--config msan` - Memory sanitizer

* `--config ubsan` - Undefined behavior sanitizer

For example to run the tests with the address sanitizer enabled:

```
bazel test -c dbg --config=asan  spectator_test --test_output=all
```

+++ test/publisher_test.cc
@@ -67,7 +67,7 @@ std::pair<int, payload_entry> get_entry(const std::vector<std::string>& strings,
 
 std::vector<payload_entry> payload_to_entries(
     const rapidjson::Document& payload) {
-  auto num_strings = payload[0].GetInt();
+  auto num_strings = static_cast<std::size_t>(payload[0].GetInt());
   std::vector<std::string> strings;
   strings.resize(num_strings);
   for (auto i = 1; i <= num_strings; ++i) {
